feat: Add blockchain output scanning functionality
Implements ability to scan Monero blockchain blocks to detect outputs
belonging to the wallet. Enables wallets to discover received funds and
maintain accurate balance by processing blocks from the network.

Includes subaddress management, reorganization handling, and support for
both full and view-only wallets.

+++ src/types.rs
@@ -173,8 +173,8 @@ pub struct SerializableOutput {
     /// Transaction hash
     pub tx_hash: [u8; 32],
 
-    /// Output index in the transaction
-    pub output_index: u8,
+    /// Output index in the transaction (u64 to support large transactions)
+    pub output_index: u64,
 
     /// Amount in atomic units (piconeros)
     pub amount: u64,
